requirements.txt
apache-airflow==2.7.0
yfinance==0.2.18
boto3==1.28.57
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
moto==4.2.5
PyYAML==6.0.1